refactor: ajustar controladores principales y entidades CustomerOrder, OrdersDetail y Product

+++ src/main/java/com/libreria_pedidos/libreria_api/model/OrdersDetail.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -16,9 +17,9 @@ public class OrdersDetail {
 
     private Integer quantity;
 
-    private double price;
+    private BigDecimal price;
 
-    private double subtotal;
+    private BigDecimal subtotal;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "product_id")
